@@ -1,25 +1,18 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="data\**" />
-    <Content Remove="data\**" />
-    <EmbeddedResource Remove="data\**" />
-    <None Remove="data\**" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="IronPython" Version="2.7.10" />
-    <PackageReference Include="Jurassic" Version="3.1.0" />
-    <PackageReference Include="Microsoft.CodeAnalysis.Scripting" Version="3.6.0" />
+    <PackageReference Include="IronPython" Version="2.7.12" />
+		<PackageReference Include="Jurassic" Version="3.2.4" />
+    <PackageReference Include="Microsoft.CodeAnalysis.Scripting" Version="4.0.1" />
     <PackageReference Include="MoonSharp" Version="2.0.0" />
     <PackageReference Include="Najlot.Log.Extensions.Logging" Version="4.0.0-alpha-6" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-    <PackageReference Include="Peachpie.AspNetCore.Web" Version="0.9.990" />
-    <PackageReference Include="Peachpie.Library.Scripting" Version="0.9.990" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="Peachpie.AspNetCore.Web" Version="1.0.11" />
+    <PackageReference Include="Peachpie.Library.Scripting" Version="1.0.11" />
   </ItemGroup>
 
 </Project>
