@@ -1,14 +1,14 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Peachpie.AspNetCore.Web;
 using Saigak.RequestHandler;
-using System.Buffers;
+using System;
 using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace Saigak
 {
-	public class Globals
+	public sealed class Globals : IDisposable
 	{
 		private static readonly byte[] _newLineBytes = new byte[] { (byte)'\n' };
 		private readonly IRequestHandler handler;
@@ -23,54 +23,36 @@ namespace Saigak
 			PhpContext = context.GetOrCreateContext();
 		}
 
-		private readonly Encoding _encoding = Encoding.UTF8;
+		private static readonly Encoding _encoding = Encoding.UTF8;
+		private static readonly byte[] _endOfLine = Encoding.UTF8.GetBytes("\n");
+
+		private readonly MemoryStream _stream = new();
+		private long _streamOffset = 0;
 
 		public async Task WriteAsync(string s)
 		{
-			if (!Context.Response.HasStarted)
-			{
-				Context.Response.StatusCode = 200;
-				Context.Response.ContentType = "text/html; charset=utf-8";
-			}
-
-			await Context.Response.WriteAsync(s, _encoding);
+			var bytes = _encoding.GetBytes(s);
+			await _stream.WriteAsync(bytes);
 		}
 
 		public async Task WriteLineAsync(string s)
 		{
-			if (!Context.Response.HasStarted)
-			{
-				Context.Response.StatusCode = 200;
-				Context.Response.ContentType = "text/html; charset=utf-8";
-			}
-
-			await Context.Response.WriteAsync(s, _encoding);
-			await Context.Response.WriteAsync("\n", _encoding);
+			var bytes = _encoding.GetBytes(s);
+			await _stream.WriteAsync(bytes);
+			await _stream.WriteAsync(_endOfLine);
 		}
 
 		public void Write(string s)
 		{
-			if (!Context.Response.HasStarted)
-			{
-				Context.Response.StatusCode = 200;
-				Context.Response.ContentType = "text/html; charset=utf-8";
-			}
-
 			var bytes = _encoding.GetBytes(s);
-			Context.Response.BodyWriter.Write(bytes);
+			_stream.Write(bytes);
 		}
 
 		public void WriteLine(string s)
 		{
-			if (!Context.Response.HasStarted)
-			{
-				Context.Response.StatusCode = 200;
-				Context.Response.ContentType = "text/html; charset=utf-8";
-			}
-
 			var bytes = _encoding.GetBytes(s);
-			Context.Response.BodyWriter.Write(bytes);
-			Context.Response.BodyWriter.Write(_newLineBytes);
+			_stream.Write(bytes);
+			_stream.Write(_endOfLine);
 		}
 
 		public string GetRequestString()
@@ -101,6 +83,31 @@ namespace Saigak
 			Context.Request.Path = new PathString(path);
 			await handler.Handle(this);
 			Context.Request.Path = originalPath;
+		}
+
+		public async Task FlushAsync()
+		{
+			if (_stream.Position == 0)
+			{
+				return;
+			}
+
+			if (!Context.Response.HasStarted)
+			{
+				Context.Response.StatusCode = 200;
+				Context.Response.ContentType = "text/html; charset=utf-8";
+			}
+
+			_stream.Seek(_streamOffset, SeekOrigin.Begin);
+			await _stream.CopyToAsync(Context.Response.Body);
+			_streamOffset = _stream.Position;
+
+			await Context.Response.Body.FlushAsync();
+		}
+
+		public void Dispose()
+		{
+			_stream.Dispose();
 		}
 	}
 }
